+++ lib/runtime-c-api/src/instance.rs
@@ -173,6 +173,7 @@ pub unsafe extern "C" fn wasmer_instance_call(
                         tag: wasmer_value_tag::WASM_F64,
                         value: wasmer_value { F64: x },
                     },
+                    Value::V128(_) => unimplemented!("calling function with V128 parameter"),
                 };
                 results[0] = ret;
             }