+++ src/bin/wasmer.rs
@@ -316,7 +316,12 @@ fn execute_wasm(options: &Run) -> Result<(), String> {
     };
 
     if !utils::is_wasm_binary(&wasm_binary) {
-        wasm_binary = wabt::wat2wasm(wasm_binary)
+        let mut features = wabt::Features::new();
+        if options.backend == Backend::LLVM {
+            // SIMD is only supported in the LLVM backend for now
+            features.enable_simd();
+        }
+        wasm_binary = wabt::wat2wasm_with_features(wasm_binary, features)
             .map_err(|e| format!("Can't convert from wast to wasm: {:?}", e))?;
     }